WIP refactor message types
